# packages/contracts/openapi/relay.yaml
openapi: 3.1.0
info:
  title: Relay API
  version: 0.1.0
paths:
...
  /v1/me:
    get:
      operationId: getCurrentUser
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current authenticated user
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [uid, email]
                properties:
                  uid: { type: string, minLength: 1 }
                  email: { type: [string, "null"] }
        "401":
          $ref: "#/components/responses/Problem"
...
  /v1/commitments/{commitment_id}/pickup-contact:
    post:
      operationId: declarePickupContact
      security: [{ bearerAuth: [] }]
      parameters:
        - name: commitment_id
          in: path
          required: true
          schema: { type: string, minLength: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PickupContactCommand" }
      responses:
        "200":
          description: Explicit pickup decision
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PickupContactResponse" }
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
...
  /v1/google/contact-picker:
    get:
      operationId: searchContactPicker
      security: [{ bearerAuth: [] }]
      parameters:
        - name: query
          in: query
          required: true
          schema: { type: string, minLength: 2 }
      responses:
        "200":
          description: Redacted current picker session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PickerSessionView" }
        "401": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /v1/actions/{action_id}:
    get:
      operationId: getAction
      security: [{ bearerAuth: [] }]
      parameters:
        - name: action_id
          in: path
          required: true
          schema: { type: string, minLength: 1 }
      responses:
        "200":
          description: Safe action status
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ActionStatusResponse" }
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
  headers:
    CorrelationId:
      required: true
      schema: { type: string, minLength: 1 }
  responses:
    Problem:
      description: Safe API error
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    ActionState:
      type: string
      enum: [planned, awaiting_approval, authorized, dispatched, in_progress, succeeded, needs_user, retryable_failure, failed, verified, handoff_opened]
...
    ActionStatusResponse:
      type: object
      additionalProperties: false
      required: [action_id, state, retry_count, correlation_id]
      properties:
        action_id: { type: string, minLength: 1 }
        state: { $ref: "#/components/schemas/ActionState" }
        retry_count: { type: integer, minimum: 0 }
        verification_evidence: { type: [object, "null"], additionalProperties: true }
        correlation_id: { type: string, minLength: 1 }
...
    PickupContactCommand:
      type: object
      additionalProperties: false
      required: [selection, expected_version]
      properties:
        selection: { type: string, enum: [no_pickup, google_picker, manual] }
        picker_session_id: { type: [string, "null"] }
        picker_contact_index: { type: [integer, "null"], minimum: 0, maximum: 19 }
        manual_display_name: { type: [string, "null"], minLength: 1, maxLength: 200 }
        manual_phone_number: { type: [string, "null"], minLength: 7, maxLength: 32 }
        expected_version: { type: integer, minimum: 1 }
    PickupContactResponse:
      type: object
      additionalProperties: false
      required: [commitment_id, version, selection, display_name]
      properties:
        commitment_id: { type: string, minLength: 1 }
        version: { type: integer, minimum: 1 }
        selection: { type: string, enum: [no_pickup, selected] }
        display_name: { type: [string, "null"] }
...
    PickerPhoneView:
      type: object
      additionalProperties: false
      required: [last4]
      properties:
        label: { type: [string, "null"], maxLength: 100 }
        last4: { type: string, pattern: "^[0-9]{4}$" }
    PickerContactView:
      type: object
      additionalProperties: false
      required: [display_name, phones]
      properties:
        display_name: { type: string, minLength: 1, maxLength: 200 }
        phones: { type: array, items: { $ref: "#/components/schemas/PickerPhoneView" } }
    PickerSessionView:
      type: object
      additionalProperties: false
      required: [session_id, contacts]
      properties:
        session_id: { type: string, minLength: 1 }
        contacts: { type: array, items: { $ref: "#/components/schemas/PickerContactView" } }
    Problem:
      type: object
      additionalProperties: false
      required: [code, message, correlation_id]
      properties:
        code: { type: string, minLength: 1 }
        message: { type: string, minLength: 1 }
        correlation_id: { type: string, minLength: 1 }